fix(natds-react): resolve textField dependecy cycle
