+++ packages/react/src/components/TextField/TextField.styles.tsx
@@ -1,17 +1,7 @@
 import { createUseStyles } from 'react-jss'
 import { Theme } from '@naturacosmeticos/natds-themes'
 import { TextFieldProps } from './TextField.props'
-
-export const getFeedbackTextColor = (theme: Theme) => ({ feedback, disabled }: TextFieldProps) => {
-  const defaultColor = disabled ? theme.color.lowEmphasis : theme.color.mediumEmphasis
-
-  const textColor = {
-    error: theme.color.alert,
-    success: theme.color.success
-  }
-
-  return feedback ? textColor[feedback] : defaultColor
-}
+import { getFeedbackTextColor } from '../Input/InputHelperText/InputHelperText.styles'
 
 const styles = createUseStyles((theme: Theme) => ({
   labelText: {
